legacy/requirements.txt
requests>=2.31.0
prettytable>=3.10.0
pyyaml>=6.0
